+++ src/Filter.js
@@ -2,6 +2,15 @@
 
 
 const Filter = (props) => {
+
+  function sortByDate(tasks) {
+    tasks = [...tasks].sort((a, b) => (a.creationTime < b.creationTime)? 1:-1)
+  }
+
+  function sortByName(tasks) {
+    tasks = [...tasks].sort((a,b) => (a.task > b.task)? 1:-1)
+  }
+
     return <div className={"container backdrop"}>
         <div className={"taskModal"}>
         <div className="justify-content-center">
@@ -11,7 +20,7 @@ const Filter = (props) => {
 
                   <button className={"col-4 priority-button"} type={"button"} 
                           onClick={() => {
-                            props.sortByName();
+                            sortByName(props.tasks);
                             props.toggleFilter();
                             }}>
                     Name
@@ -19,7 +28,7 @@ const Filter = (props) => {
 
                   <button className={"col-4 priority-button"} type={"button"} 
                           onClick={() => {
-                            
+                            // props.sortByPriority();
                             props.toggleFilter();
                   }}>
                     Priority
@@ -27,10 +36,10 @@ const Filter = (props) => {
 
                   <button className={"col-4 priority-button"} type={"button"} 
                           onClick={() => {
-                            props.sortByDate();
+                            sortByDate(props.tasks);
                             props.toggleFilter();
                   }}>
-                    Creation Date
+                    Date
                   </button>
 
                   </div>
