+++ src/ToggleBar.js
@@ -1,5 +1,7 @@
 import TaskList from './TaskList.js';
 import TaskAdder from './TaskAdder.js';
+import Header from './Header.js';
+import Filter from'./Filter.js';
 
 import { useState } from 'react';
 import { collection, deleteDoc, doc, getFirestore, updateDoc, query, setDoc/* , serverTimestamp */ } from "firebase/firestore";
@@ -13,6 +15,7 @@ const ToggleBar = (props) => {
   const [showAllTasks, setShowAllTasks] = useState(true);
   const displayAllTasks = () => setShowAllTasks(true);
   const hideCompletedTasks = () => setShowAllTasks(false);
+  const [showFilter, setShowFilter] = useState(false);
   const [currentTaskList, setCurrentTaskList] = useState((props.taskLists && props.taskLists.length > 0) ? props.taskLists[0].id : "");
   const q = query(collection(props.db, props.collectionName, currentTaskList, subCollectionName));
   const [tasks, loading, error] = useCollectionData(q);
@@ -33,21 +36,25 @@ const ToggleBar = (props) => {
       });
   }
 
-  function sortByDate() {
-    tasks = [...tasks].sort((a, b) => (a.creationTime < b.creationTime)? 1:-1)
-  }
+  
 
-  function sortByName() {
-    tasks = [...tasks].sort((a,b) => (a.task < b.task)? 1:-1)
+  function toggleFilter() {
+    setShowFilter(!showFilter);
   }
 
 
   if (loading) {
-    return <p>Loading</p>;
+    return (<div className="container">
+      <Header/>
+      <p>Loading</p>;
+    </div>);
   }
 
   if (error) {
-    <p>ERROR</p>
+    return (<div className="container">
+      <Header/>
+      <p>Error</p>;
+    </div>);
   }
   return (<div>
     <div className="row">
@@ -77,12 +84,17 @@ const ToggleBar = (props) => {
           </Dropdown.Menu>
         </Dropdown>
       </div>
-      <div className='col-6'onClick={props.toggleFilter}>
+      <div className='col-6'onClick={toggleFilter}>
         <div className="Tab">
           Filter
         </div>
       </div>
     </div>
+    {showFilter && <Filter toggleFilter={toggleFilter} 
+                                      taskToEdit={props.taskToEdit} 
+                                      onItemChanged={props.onItemChanged}
+                                      tasks={tasks}
+                                      />}
     <TaskAdder data={tasks} onAddTask={props.onAddTask} />
     <div className="row">
       <TaskList data={tasks}
